Added preliminary MIC support to AQ
Added preliminary MIC support to AQ.  Note that MIC doesn't send us any metadata
with our payload so we have to depend entirely on the payload from the device.
Fortunately we can fix this by utilizing the sysid to uniquely identify the devices
and look up the calibration data.

Also included the calibration data in the repository since having to maintain
an S3 bucket is not going to work as a long term solution.

Listeners now have to be explicitly enabled with one of these
command line options: --enable-mic or --enable-horde.  The UDP
listener only starts if you provide the --udp-listener with
a listen address.

Changed the miclistener code a bit so it takes a config object.