Populate DeviceID from calibration data when missing from datapoint.
This is a workaround for when we use the MIC listener.  The device ID
gets lost because MIC doesn't propagate the metadata from the IoT-GW.

This change also makes sure we can't start the server if a MIC
listener is active but no credentials have been entered into
the right environment variables.