Changed the way calibration data is handled.
Rather than distributing the calibration data via AWS S3 we will now
be distributing it by putting it in Github.  This means that anyone operating
an AQ server has to be a bit on their toes about updating the data themselves
but it removes the burden of maintaining the calibration data from Telenor.

On startup the AQ server will look for calibration data in the directory
specified by the -m or --cal-data-dir command line options.  By default this
is the "calibration-data" directory relative to current directory.  This is
where the data is when the code is checked out and where we will check in
new calibration data (for now).

Usually the calibration data will be in the database, so on startup it will
just traverse this directory and look for any new data.  If it finds data
that isn't already in the database it will add it.